Allows lxc config file customization
This patch makes it possible to customize the created lxc config file
independently for every host. For example, if the hostname is "sample"
and a file /var/lib/lxc/sample.config exists, the content of this file
will be added to /var/lib/lxc/sample/config.

The behavior is unmodified if the file does not exist (except for an
added newline character).